Latest Changes

+++ src/test/jmeter/GrafanaServer.java
@@ -6,6 +6,7 @@ import org.testng.annotations.Test;
 import java.io.BufferedReader;
 import java.io.InputStreamReader;
 import java.time.Duration;
+import java.util.concurrent.TimeUnit;
 
 public class GrafanaServer {
 
@@ -19,21 +20,30 @@ public class GrafanaServer {
             processBuilder.redirectErrorStream(true);
             Process process = processBuilder.start();
             log.info("Grafana Started");
-            BufferedReader reader = new BufferedReader(new InputStreamReader(process.getInputStream()));
-            while (!SharedStatus.jmxExecutionCompleted.get()) {
-                try{
-                    String line = reader.readLine();
-                    if(line != null){
-                        System.out.println("Grafana: "+ line);
+            //SharedStatus.jmxCompletedLatch.await();
+            try {
+                BufferedReader reader = new BufferedReader(new InputStreamReader(process.getInputStream()));
+                while (!SharedStatus.jmxCompletedLatch.await(1, TimeUnit.SECONDS)) {
+                    while (reader.ready()) {
+                        String line = reader.readLine();
+                        if (line != null && !line.isBlank() && line.contains("HTTP Server Listen")) {
+                            System.out.println("Grafana: " + line);
+                        }
                     }
-                } catch (Exception e) {
-                    log.info("Error reading output of Grafana : {}", e.getMessage());
+                }
+            } catch (Exception e) {
+                throw new RuntimeException(e);
+            }
+            log.info("Waiting 5 mins after jmeter execution done");
+            Thread.sleep(Duration.ofMinutes(1));
+            log.info("Stopping Grafana");
+            if (process.isAlive()) {
+                process.destroy();
+                process.waitFor(10, TimeUnit.SECONDS);
+                if (process.isAlive()) {
+                    process.destroyForcibly();
                 }
             }
-            Thread.sleep(Duration.ofMinutes(1));
-            log.info("Waiting 5 mins after jmeter execution done");
-            log.info("Stopping Grafana");
-            process.destroyForcibly();
             Process process1 = new ProcessBuilder("taskkill", "/F", "/IM", "grafana.exe").start();
             process1.waitFor();
             log.info("Grafana Server Stopped successfully");
